Move and rename map view
It was weird to have it in a sub-dir and then creating a view higher up
in the hierarchy, also keep the var name (EditorMapView) to match it’s
file path which is more intutitive
