+++ lib/assets/javascripts/cartodb3/editor.js
@@ -15,7 +15,7 @@ var _ = require('underscore');
 var cdb = require('cartodb.js');
 var deepInsights = require('cartodb-deep-insights.js');
 var ConfigModel = require('./data/config-model');
-var EditorMapView = require('./editor/map/map-view');
+var EditorMapView = require('./editor/editor-map-view');
 var AnalysisDefinitionNodesCollection = require('./data/analysis-definition-nodes-collection');
 var AnalysisDefinitionsCollection = require('./data/analysis-definitions-collection');
 var LayerDefinitionsCollection = require('./data/layer-definitions-collection');
