+++ lib/assets/javascripts/cartodb3/editor/editor-map-view.js
@@ -1,10 +1,10 @@
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
-var EditorView = require('../editor-view');
-var AddAnalysisView = require('../../components/modals/add-analysis/add-analysis-view');
-var StackLayoutView = require('../../components/stack-layout/stack-layout-view');
-var LayerContentView = require('../layers/layer-content-view');
-var WidgetsFormContentView = require('../widgets/widgets-form/widgets-form-content-view');
+var EditorView = require('./editor-view');
+var AddAnalysisView = require('../components/modals/add-analysis/add-analysis-view');
+var StackLayoutView = require('../components/stack-layout/stack-layout-view');
+var LayerContentView = require('./layers/layer-content-view');
+var WidgetsFormContentView = require('./widgets/widgets-form/widgets-form-content-view');
 
 module.exports = cdb.core.View.extend({
   className: 'MapEditor',
