+++ lib/assets/test/spec/cartodb3/editor/editor-map-view.spec.js
@@ -2,13 +2,13 @@ var _ = require('underscore');
 var $ = require('jquery');
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
-var LayerDefinitionModel = require('../../../../../javascripts/cartodb3/data/layer-definition-model');
-var AddAnalysisView = require('../../../../../javascripts/cartodb3/components/modals/add-analysis/add-analysis-view');
-var ModalsService = require('../../../../../javascripts/cartodb3/components/modals/modals-service-model');
-var EditorMapView = require('../../../../../javascripts/cartodb3/editor/map/map-view');
-var EditorModel = require('../../../../../javascripts/cartodb3/data/editor-model');
+var LayerDefinitionModel = require('../../../../javascripts/cartodb3/data/layer-definition-model');
+var AddAnalysisView = require('../../../../javascripts/cartodb3/components/modals/add-analysis/add-analysis-view');
+var ModalsService = require('../../../../javascripts/cartodb3/components/modals/modals-service-model');
+var EditorMapView = require('../../../../javascripts/cartodb3/editor/editor-map-view');
+var EditorModel = require('../../../../javascripts/cartodb3/data/editor-model');
 
-describe('editor/map/map-view', function () {
+describe('editor/editor-map-view', function () {
   beforeEach(function () {
     var configModel = 'c';
     var basemapLayerDefModel = new LayerDefinitionModel({
